Refactor the IR representation to be reference-based
This is necessary since a parent pointer, which references
a specific expression by definition, does not make sense
in a value-based IR representation. Refactoring the code
wasn't actually that bad and enabled us to address some
long-standing problems.

In addition, this commit implements the actual parent pointer
and the cursor class.

Note: This change breaks the loop optimizer. Since we will
          refactor the loop optimizer infrastructure anyway within
          the next weeks, I will simply disable it for now.