Derive a tighter upper bound for non-unit stride loops during axiom analysis
If a loop has non-unit stride, the induction variable will not
run through all values between the bounds and might actually
stop much earlier then indicated by our current estimate.

Currently, we don't exploit that the induction variable
does not run through all values in that range, but we might
want to add a corresponding axiom in the future.