Add support infrastructure for the construction of affine constraints
Since affine constraints (and affine expressions) are the foundation of
many important analysis and optimization passes it makes sense to
transfer this functionality into a generic set of helper classes and functions.

This functionality is currently *not* tested. I will use it in the next commit to
implement the value set analysis pass which will stress the next infrastructure.