Implement the value set analysis pass
This pass tries to determine the set of possible values for an expression
within the context of a certain task. It is guaranteed that the obtained set
 is a superset of the actual value set of the analyzed expression.

I made sure that the analysis is quite accurate since its results are very useful during a
wide range of other analyses and optimizations (e.g. alias analysis and memory
optimizations).