Add an advanced alias analysis to tell apart accesses to different portions of an array
If encountering array accesses to two potentially aliasing objects, the new analysis
calculates the value sets of all indices. If the value sets of the corresponding
indices do not overlap, we can be sure that the accesses can never alias.

With the new path-based alias analysis framework, we get much more precise results
than previously but I am not convinced that the structure is as good as it could be.
Most importantly, I think the framework is not as composable as it could be.

Anyway, the new implementation is a clear improvement and ready to ship
so lets do just this for the time being.

In addition, this commit makes the use of the basic alias analysis in the task invariants
analysis more consistent.