Remove the old array analysis
The old array analysis has served us wll so far but the new framework is much more
powerful and generic. The only missing piece is to wire the code generator
up to the new framework. The essential missing part is the reorganization
of the code generator into a transformation pass so it can access the analysis
properly.