Implement the static schedule analysis pass
The analysis tries to find parts of the schedule given by the IR  whose structure is
completely determined at compile time (thus static).

The static structure allows us to reason about these parts in a much
more elaborate manner. For example, to prove that a certain reordering
of tasks is valid is much easier in this representation.  As a side effect,
we can easily prove that a set of tasks can be executed concurrently
(All permutations are valid).