Add the first parts of the performance modelling infrastructure
In addition to the interface, this commit also adds two implementations.

The first implementation is a simple statistical model based on Boost Accumulators
which uses the moments of a recorded frequency distribution for every kernel instance.

The second implementation, the unified performance model, provides an abstraction
over such concrete models and will allow us to pick a specific model based on
the properties of a certain kernel. Specifically, I will add an advanced performance model
in a followup commit for kernels with scalar parameters which will exploit this
infrastructure.