Change the type of the instruction and stack pointer to signed integer
Modulo arithmetic is the root of all evil.

In addition, this commit adds some detail improvements.

I will stop optimizing the bytecode interpreter right here.
I do not think that we will obtain a much better performance
using the current execution model. The execution time of the base
operations is already in the order of only a few nanoseconds.

If somebody wants to further improve the performance, the next
step should be the implementation and testing of a register machine.
Currently, most of the execution time is spent accessing the stack
so the usage of registers should speed up the execution. At least
if the compiler manages to generate efficient code for the register
machine.

Surprisingly, at least for me, the AST interpreter performs quite well.
The execution times are similar to the bytecode interpreter and
are only significantly larger in some corner cases. It might be
a good idea to reevaluate this in the future and reenable
the AST interpreter if it is beneficial.