Add a performance model based on symbolic regression
The new model tries to find a suitable model function over time which
characterizes the measured data. These models only depend
on the scalar parameters of the kernel. Therefore, the
overall performance model dispatches each operation
to a partial model for each combination of non-scalar parameters
to cover all cases.

Currently, the model might consume a huge chunk of memory
to store the data and the models. To avoid this, we have to
restrict the amount of stored data to a sufficiently small value.
Since this implies that we would need to discard some data eventually,
we need to come forward with a good idea to categorize the data
according to its importance before we actually implement such
a limit.

Since this is a more general problem which affects all performance
models to some extent, the solution should also be generic enough
to work for any performance model.