Restrict the number of collected data points for symbolic regressions
Previously, I had planned to give the user the choice if he wants to
restrict the number of points or not, but it turned out to be
too bothersome to implement given the performance constraints.
Since the infrastructure is already quite specialized and
we will never collected an infinite amount of data for performance
modelling, this shouldn't be a big loss. As memory is only
allocated as the dataset grows, one could obtain a practically
unrestricted regression by simply setting the window size to
a sufficiently large value.