Move the component resolution into the address space implementation
In general, the user-side wrapper can't do this, since
it might not have access to the memory location in which the
corresponding page resides.

This commit also ensures that all referenced pages stay resident.
This has the unfortunate side effect that we can not drop pages
of subobjects as long as the parent object is still resident which
could lock down a large chunk of memory.

I am not sure if this will be a problem in practice since we rarely
materialize aggregate objects anyway, and those are usually fairly
small. We could also counteract this problem with a smarter eviction
algorithm which will drop all unused aggregate pages before trying to
evict actual data.