Move the data flow graph construction to the runtime
Previously, each backend was constructing the data flow graph by itself.
Not only did this place a huge implementation cost onto each backend,
but it also lead to suboptimal scheduling decision because the necessary information
was only available in the last stages of the scheduling pipeline.

This commit moves the entire data flow graph handling into the runtime so that
the entire pipeline can use this information.

To achieve this goal, a huge refactoring of our previous framework was necessary.
On the upside, this allowed us to remove a large amount of cruft and to streamline
the implementation. While I am still not completely happy with the resulting
implementation, it is a definite improvement since most of the runtime (and
the backends) can now be completely oblivious to the meaning of the data flow
graph.

The largest part of the refactoring goes into the new distributed_future type and
associated classes which encapsulate most of the old logic in a reusable way.
A distributed_future is a variant of future which can be passed between different
localities which is not possible using HPX's normal future. The data flow graph
framework mostly uses it as a bridge between chains of futures on different localities
which represent the logic of the data flow graph. In particular, we can now pass futures
(and by association the dependency information represented by it) down the scheduling
pipeline.

Currently only the runtime needs to construct new tasks and thus
manipulate the data flow graph. Later on, certain VPUs might want to
start new subtasks. In the long run, we should certainly factor the logic
out of the runtime into a reusable infrastructure.