Implement a facility for using C++ code as a kernel
This commit revives the old foreign kernel facilities and
finalizes them.

To support multiple devices, one can register different versions
of the kernel each targeting a different architecture.

In principle, this facility works similar to the view feature
except that a foreign kernel will be scheduled by Qubus
in the same way as any other computation. This makes it
especially useful for heavy computations which one does not want to
implement using Qubus IR directly for one reason or another.
For exmaple porting legacy code to Qubus bit by bit might be
a good application.