Add the ability to get a view for already locked objects
This makes it possible to fix the deadlock in the regression-based
performance model since we can now access the contents of an already
locked object without trying to lock it again.

This commit also adds a test case triggering the deadlock.

This fixes issue #61.

In the long run, we should make this type-safe by using another type
for locked objects.

Fly-by changes: Fix some minor bugs.