Fix the variable access analysis
Now the analysis actually does what is supposed to do, namely
to enumerate *all* accesses to variables which are defined outside
of the specified task.

A followup commit will add the ability to enumerate all accesses
including ones which reference internal variables of the task.

In addition, this commit adds a unit test for the pass.