Aggregate all used LLVM libraries into a shared library
This ensures that all global variables of LLVM (like the
command-line options) are only initialized once independent
of the configuration of LLVM. Previously Qubus would die if more than
one backend (using LLVM) was loaded since some options were initialized
twice.

I don't really like that solution since it requires us to know each
used component of LLVM beforehand but so far I have not found a better
solution. In principle, we have the same problem if a user also links
LLVM through his own code or another library. But in this case we
really can not do anything but to tell the user to link LLVM
dynamically in both cases.

Also return the generated function after compiling the entry
function of a kernel to make it easier to manipulated the function
afterwards.