Split the backend scan into two phases (host + VPU)
Currently, Qubus' architecture assumes that there is a portion of
the hardware (processor + memory) that can communicate and control
every other part of the system. This host system, mapped by a
host backend, is the lingua franca of Qubus. Due to this, a host
backend has to expose additional capabilities compared to a
regular backend.

This has always been that case, but now we acknowledge this fact by
splitting the backend scan in a host part, which chooses exactly **one**
host backend, and a second pass for the regular/VPU backends.

This simplfies some previously awkward code and is necessary for the
followup commit which will introduce a communication channel between
the host backend and any regular backend.