Add a new compiler hook for the root skeleton
The new hook is intended for use cases where the root of
the skeleton tree has to be handled in a different fashion,
for example if prolog or epilog code needs to be emitted.

The default implementation simply forwards the code to the
generic compilation routine.

To support this use case, this commit also changes the compilation
of the entry point such that the body is emitted directly instead
of using another function to hide the body.