Emit correct prologue code for the PTX kernel
If the first expression is an unordered loop,
we will emit a grid-stride loop corresponding to that loop.
Otherwise we will serialize the execution.

This is not very efficient in most cases but gets us running
and we can improve the generated code in the future.

The biggest remaining problem in that regard is that CUDA
does not support synchronization across the entire grid which
makes it necessary to perform some of the operations on the host
processor (or another auxiliary processor). This is not trivial
with our current backend architecture and needs some consideration.