Reimplement dense_hash_map
The new implementation is based on Malte Skarupke's fantastic flat
hash map (https://github.com/skarupke/flat_hash_map).

In addition, this commit adds a corresponding hash set.

The new implementation has the great benefit that it does not need
special keys and can thus be used as a true drop-in replacement of
the standard maps or sets.

This resolves issue #52.