Fix the prefix determination algorithm
Previously, we were trying to load the core library itself to determine
the prefix. This breaks, if the dynamic linker can not find
the library in her own context. Most notably, this can happen if
the core library is found through the RUNTIME_PATH property of her
dependent. Since RUNTIME_PATH is not transitive, it is possible
that the core library can not find itself in this case.

This commit fixes that situation by hard coding the install prefix
of Qubus into the application. To preserve the current behavior that
Qubus can be used directly from the build tree, we use an alternative
version of the code in this case.

Since the prefix is now set to a static value, we loose the ability
to move the framework after installation. In my opinion, we do not
loose much. In this case, the user needs to intervene anyway since
he has to ensure that the core library can find itself. If we decide
that we need this feature later on, we could simply add a flag
to enable the old "portable" mode. This would give us the opportunity
to make the user aware of the necessary precautions and ensure that
he is not surprised by the subtle changes in behavior.