Tweak the LTO setup
Not all compilers support LTO if compiling with debug symbols
so we only enable it in release mode. Since our primary use case
is to further optimize the release build, this commit simply disables
LTO for all compilers.

In addition, this commit disables LTO for all static libraries for now,
since these might be propagated to an end user target and break non-LTO
builds of these targets.