Revise the object model
This commit includes the following major changes:

* Add a standardized textual represenation of the IR.
* Add modules as the largest unit of code.
* Add support for fully relocatable structs
* Unify arrays and array slices
* Add support for fuzz testing using libFuzzer
* Greatly simplify the data dependency tracking in the frontend

Currently, the code for sparse tensors is not working and I have
disabled the corresponding tests. To define the sparse tensor
type for arbitrary value types and ranks, we need a language feature
similar to C++ templates. Since this commit is already too large
as is and contains many significant improvements, I have decided
to move this feature to a future MR.