Split the runtime library into a core part and the actual runtime
This is necessary since the backend libraries depend on some
portions of the current runtime library while the runtime itself
depends on them forming a cycle in the dependency graph. To avoid
this, all parts which might be used by other components are separated
into their own library. Using this structure, both the runtime
and the backends can depend on these core components which should
avoid any conflicts during linking.